Updated some models and tests.
