+++ src/common/models/index.ts
@@ -1,0 +1,25 @@
+import { Issue } from './issue';
+import { Contact } from './contact';
+import { ContactList } from './contactList';
+import {
+  Outcome,
+  OutcomeButton,
+  LocationFetchType,
+  FieldOffice,
+  CategoryMap,
+  GeolocationPosition,
+  CountData
+} from './model';
+
+export {
+  Issue,
+  Contact,
+  ContactList,
+  Outcome,
+  OutcomeButton,
+  LocationFetchType,
+  FieldOffice,
+  CategoryMap,
+  GeolocationPosition,
+  CountData
+};
