+++ src/components/call/CallHeader.test.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
+
 import { CallHeader } from './index';
-import { Issue } from '../../common/model';
+import { Issue } from '../../common/models';
 
 test('Call header component should be rendered if passed a valid object', () => {
   const issue: Issue = Object.assign({}, new Issue(), {
