+++ src/components/call/ContactDetails.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Issue, Contact } from '../../common/model';
+import { Issue, Contact } from '../../common/models';
 import { ContactOffices } from './index';
 import { makePhoneLink } from '../shared/jsxUtils';
