+++ src/components/call/Script.test.tsx
@@ -4,34 +4,25 @@ import { shallow } from 'enzyme';
 import i18n from '../../services/i18n';
 import { Script } from './';
 import { getContactNameWithTitle } from './Script';
-import {
-  Contact,
-  Issue,
-  LocationUiState,
-  LocationFetchType
-} from '../../common/model';
 import { LocationState } from '../../redux/location';
 import { linkRefRenderer } from '../shared/markdown-utils';
+import { Issue, mockIssue } from '../../common/models/issue';
+import { Contact, LocationFetchType } from '../../common/models';
+import { mockContact } from '../../common/models/contact';
 
 test('Script component should be rendered if passed a valid object', () => {
-  const issue: Issue = Object.assign({}, new Issue(), {
-    id: '1',
-    name: 'testName'
-  });
   const locationState: LocationState = {
     address: '1234',
     cachedCity: 'Anytown',
     splitDistrict: false,
     invalidAddress: false,
-    uiState: LocationUiState.LOCATION_FOUND,
     locationFetchType: LocationFetchType.BROWSER_GEOLOCATION
   };
   const component = shallow(
     <Script
-      issue={issue}
-      contactIndex={0}
+      issue={mockIssue}
+      currentContact={mockContact}
       locationState={locationState}
-      t={i18n.t}
     />
   );
   expect(component).toMatchSnapshot();
@@ -39,39 +30,19 @@ test('Script component should be rendered if passed a valid object', () => {
 
 describe('when the script text is shown', () => {
   test('it uses ReactMarkdown for display formatting', () => {
-    const t = jest.fn().mockReturnValue('');
-    const contacts: Contact[] = [
-      {
-        id: '293',
-        area: 'House',
-        name: 'Googly Moogly',
-        phone: '',
-        state: '',
-        party: 'Independent',
-        reason: ''
-      }
-    ];
-    const issue: Issue = Object.assign({}, new Issue(), {
-      id: '1',
-      name: 'testName',
-      contacts: contacts,
-      script: 'script_text'
-    });
     const locationState: LocationState = {
       address: '1234',
       cachedCity: 'Anytown',
       splitDistrict: false,
       invalidAddress: false,
-      uiState: LocationUiState.LOCATION_FOUND,
       locationFetchType: LocationFetchType.BROWSER_GEOLOCATION
     };
 
     const wrapper = shallow(
       <Script
-        issue={issue}
-        contactIndex={0}
+        issue={mockIssue}
+        currentContact={mockContact}
         locationState={locationState}
-        t={t}
       />
     );
     expect(
@@ -88,7 +59,7 @@ describe('when the script text is shown', () => {
 
 describe('getContactNameWithTitle', () => {
   const contacts: Contact[] = [
-    {
+    Object.assign(new Contact(), {
       id: '293',
       area: 'House',
       name: 'Googly Moogly',
@@ -96,8 +67,8 @@ describe('getContactNameWithTitle', () => {
       state: '',
       party: '',
       reason: ''
-    },
-    {
+    }),
+    Object.assign(new Contact(), {
       id: '666',
       area: 'Senate',
       name: 'Damian Longhoof',
@@ -105,8 +76,8 @@ describe('getContactNameWithTitle', () => {
       state: '',
       party: '',
       reason: ''
-    },
-    {
+    }),
+    Object.assign(new Contact(), {
       id: '123',
       area: 'AttorneyGeneral',
       name: 'Betty White',
@@ -114,8 +85,8 @@ describe('getContactNameWithTitle', () => {
       state: '',
       party: '',
       reason: ''
-    },
-    {
+    }),
+    Object.assign(new Contact(), {
       id: '789',
       area: 'StateLower',
       name: 'Luna Lovegood',
@@ -123,8 +94,8 @@ describe('getContactNameWithTitle', () => {
       state: '',
       party: '',
       reason: ''
-    },
-    {
+    }),
+    Object.assign(new Contact(), {
       id: '739',
       area: 'Underground',
       name: 'Lisbeth Salander',
@@ -132,20 +103,20 @@ describe('getContactNameWithTitle', () => {
       state: '',
       party: '',
       reason: ''
-    }
+    })
   ];
   it('returns the name of the contact with their title', () => {
-    const repTitle = getContactNameWithTitle(contacts, 0);
+    const repTitle = getContactNameWithTitle(contacts[0]);
     expect(repTitle).toBe('Rep. Googly Moogly');
-    const senateTitle = getContactNameWithTitle(contacts, 1);
+    const senateTitle = getContactNameWithTitle(contacts[1]);
     expect(senateTitle).toBe('Senator Damian Longhoof');
-    const agTitle = getContactNameWithTitle(contacts, 2);
+    const agTitle = getContactNameWithTitle(contacts[2]);
     expect(agTitle).toBe('Attorney General Betty White');
-    const legTitle = getContactNameWithTitle(contacts, 3);
+    const legTitle = getContactNameWithTitle(contacts[3]);
     expect(legTitle).toBe('Legislator Luna Lovegood');
   });
   it('returns just the name of the contact when the title is unknown', () => {
-    const unknownTitle = getContactNameWithTitle(contacts, 4);
+    const unknownTitle = getContactNameWithTitle(contacts[4]);
     expect(unknownTitle).toBe('Lisbeth Salander');
   });
 });
