+++ src/components/issues/IssuesListItem.test.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { shallow } from 'enzyme';
+
 import { IssuesListItem } from './index';
-import { Issue } from '../../common/model';
+import { Issue } from '../../common/models';
 
 test('IssuesListItem issue click passes issue id', () => {
   const id = 'id';
