+++ src/components/layout/Layout.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 
-import { Issue } from '../../common/model';
+import { Issue } from '../../common/models';
 import { SidebarHeader, Sidebar, Footer, Header } from './index';
 
 import {
