+++ src/components/location/Location.test.tsx
@@ -3,7 +3,7 @@ import { shallow } from 'enzyme';
 import i18n from '../../services/i18n';
 import { Location } from './index';
 import { LocationState } from '../../redux/location';
-import { LocationFetchType, LocationUiState } from '../../common/model';
+import { LocationFetchType } from '../../common/models';
 
 test('Location component should show location prop value if locationState.address is defined', () => {
   const locationState: LocationState = {
@@ -11,13 +11,10 @@ test('Location component should show location prop value if locationState.addres
     cachedCity: '',
     splitDistrict: false,
     invalidAddress: false,
-    uiState: LocationUiState.LOCATION_FOUND,
     locationFetchType: LocationFetchType.BROWSER_GEOLOCATION
   };
 
-  const component = shallow(
-    <Location t={i18n.t} locationState={locationState} />
-  );
+  const component = shallow(<Location locationState={locationState} />);
   const node = component.find('#locationMessage span');
   expect(node.text()).toEqual(locationState.address);
 });
@@ -28,13 +25,10 @@ test('Location component should show location prop value if locationState.cached
     cachedCity: 'Cached Address',
     splitDistrict: false,
     invalidAddress: false,
-    uiState: LocationUiState.LOCATION_FOUND,
     locationFetchType: LocationFetchType.BROWSER_GEOLOCATION
   };
 
-  const component = shallow(
-    <Location t={i18n.t} locationState={locationState} />
-  );
+  const component = shallow(<Location locationState={locationState} />);
   const node = component.find('#locationMessage span');
   expect(node.text()).toEqual(locationState.cachedCity);
 });
@@ -45,13 +39,10 @@ test('Should show "Getting your location" label if fetching location', () => {
     cachedCity: '',
     splitDistrict: false,
     invalidAddress: false,
-    uiState: LocationUiState.FETCHING_LOCATION,
     locationFetchType: LocationFetchType.CACHED_ADDRESS
   };
 
-  const component = shallow(
-    <Location locationState={locationState} t={i18n.t} />
-  );
+  const component = shallow(<Location locationState={locationState} />);
   const label = component.find('p.loadingAnimation').first();
   expect(label).toBeDefined();
 });
@@ -62,13 +53,10 @@ test('If address is invalid, show proper message and form with input and "Go" bu
     cachedCity: '',
     splitDistrict: false,
     invalidAddress: false,
-    uiState: LocationUiState.LOCATION_ERROR,
     locationFetchType: LocationFetchType.CACHED_ADDRESS
   };
 
-  const component = shallow(
-    <Location locationState={locationState} t={i18n.t} />
-  );
+  const component = shallow(<Location locationState={locationState} />);
   const label = component.find('p[role="alert"]');
   expect(label).toBeDefined();
   const input = component.find('form input');
