+++ src/components/shared/utils.ts
@@ -1,5 +1,5 @@
 import * as Constants from '../../common/constants';
-import { Issue } from './../../common/model';
+import { Issue } from './../../common/models';
 import { RemoteDataState } from '../../redux/remoteData';
 
 import { find } from 'lodash';
