+++ src/redux/remoteData/asyncActionCreator.test.ts
@@ -3,16 +3,12 @@ import configureStore from 'redux-mock-store';
 import * as moxios from 'moxios';
 
 import { RemoteDataActionType } from './action';
-import { fetchCallCount, fetchAllIssues, fetchLocationByIP } from './index';
+import { fetchCallCount } from './index';
 import { ApplicationState } from './../root';
-import {
-  ApiData,
-  Issue,
-  IpInfoData,
-  LocationFetchType,
-  LocationUiState
-} from './../../common/model';
+import { Issue, LocationFetchType } from './../../common/models';
 import * as Constants from '../../common/constants';
+import { IssueData } from '../../common/models/model';
+import { getAllIssues } from '../../services/apiServices';
 
 const middlewares = [thunk];
 const mockStore = configureStore(middlewares);
@@ -28,7 +24,7 @@ afterEach(() => {
 test('getApiData() action creator functions correctly', () => {
   const address = 'New Gloucester, ME';
   const issueName = 'Issue';
-  const apiData: ApiData = getApiDataResponse(address, issueName);
+  const apiData: IssueData = getApiDataResponse(address, issueName);
   moxios.stubRequest(
     `${Constants.ISSUES_API_URL}${encodeURIComponent(address)}`,
     { response: apiData }
@@ -40,13 +36,12 @@ test('getApiData() action creator functions correctly', () => {
     cachedCity: '',
     splitDistrict: false,
     invalidAddress: false,
-    uiState: LocationUiState.FETCHING_LOCATION,
     locationFetchType: LocationFetchType.CACHED_ADDRESS
   };
   initialState.locationState = locationState;
   const store = mockStore(initialState);
   // tslint:disable-next-line:no-any
-  store.dispatch<any>(fetchAllIssues(address)).then(() => {
+  store.dispatch<any>(getAllIssues()).then(() => {
     const actions = store.getActions();
     // console.log('Actions', actions);
     expect(actions[1].payload).toEqual(address);
@@ -54,42 +49,14 @@ test('getApiData() action creator functions correctly', () => {
   });
 });
 
-const getApiDataResponse = (address, issueName): ApiData => {
+const getApiDataResponse = (address, issueName): IssueData => {
   const mockIssue = Object.assign({}, new Issue(), { name: issueName });
 
-  const mockResponse: ApiData = {
-    splitDistrict: false,
-    invalidAddress: false,
-    normalizedLocation: address,
-    divisions: [],
+  const mockResponse: IssueData = {
     issues: [mockIssue]
   };
   return mockResponse;
 };
-
-// FIME: This throws:
-//   TypeError: index_1.fetchLocationByIP is not a function
-test.skip('fetchLocationByIP() action creator works correctly', () => {
-  const data: IpInfoData = {
-    city: 'New City',
-    country: 'USA',
-    hostname: 'foobar.com',
-    org: '5 Calls',
-    postal: '04260',
-    loc: '-43.00, -70.00',
-    ip: '127.0.0.1',
-    region: 'New England'
-  };
-  moxios.stubRequest(/json/, { response: data });
-  const initialState = {} as ApplicationState;
-  // initialState.locationState = {address: ''};
-  const store = mockStore(initialState);
-  // tslint:disable-next-line:no-any
-  store.dispatch<any>(fetchLocationByIP()).then(() => {
-    // const actions = store.getActions();
-    // console.log('fetchLocationByIP() Actions', actions);
-  });
-});
 
 test('fetchCallCount() action creator dispatches proper action', () => {
   const count = 999999;
